+++ src/app/system/external-services/notification/notification.resolver.ts
@@ -1,6 +1,7 @@
 /** Angular Imports */
 import { Injectable } from '@angular/core';
 import { Resolve } from '@angular/router';
+import { SettingsService } from 'app/settings/settings.service';
 
 /** rxjs Imports */
 import { Observable } from 'rxjs';
@@ -17,14 +18,14 @@ export class NotificationConfigurationResolver implements Resolve<Object> {
   /**
    * @param {SystemService} systemService System service.
    */
-  constructor(private systemService: SystemService) {}
+  constructor(private systemService: SystemService, private settingsService: SettingsService ) {}
 
   /**
    * Returns the Notification Configuration data.
    * @returns {Observable<any>}
    */
   resolve(): Observable<any> {
-    return this.systemService.getExternalConfiguration('NOTIFICATION');
+    return this.systemService.searchExternalConfiguration('NOTIFICATION', null, null);
   }
 
 }