+++ src/app/system/system-routing.module.ts
@@ -93,6 +93,7 @@ import { AddPaymentProviderComponent } from './external-services/payment-provide
 import { EditPaymentProviderComponent } from './external-services/payment-provider/edit-payment-provider/edit-payment-provider.component';
 import { OrderIntegrationComponent } from './external-services/order-integration/order-integration.component';
 import { EditOrderIntegrationComponent } from './external-services/order-integration/edit-order-integration/edit-order-integration.component';
+import { AddNotificationComponent } from './external-services/notification/add-notification/add-notification.component';
 
 const routes: Routes = [
   Route.withShell([
@@ -238,9 +239,11 @@ const routes: Routes = [
                   path: 'edit',
                   component: EditNotificationComponent,
                   data: { title: extract('labels.text.Edit Notification Configuration'), breadcrumb: 'Edit' },
-                  resolve: {
-                    notificationConfiguration: NotificationConfigurationResolver,
-                  },
+                },
+                {
+                  path: 'add',
+                  component: AddNotificationComponent,
+                  data: { title: extract('labels.text.Add Notification Configuration'), breadcrumb: 'Add' }
                 },
               ],
             },