+++ RELEASE_NOTES.md
@@ -5,6 +5,22 @@
         * [CP-3666] - Refactor Template Configuration with HTML Support Per Country
     * Navigation Bar
         * [CP-4004] - Disabling the notification tray for now since it's not being consumed   
+## Version 1.4.4 - Community 1.0.0
+    * Clients
+        * [CP-3935] - Add time a client landed in the checker inbox
+        * [CP-3971] - Rename phone number field in UI to Paid By in transaction history
+        * [CP-3536] - Fix Cannot read properties of undefined errors reported by Sentry
+
+## Version 1.4.3 - Community 1.0.0
+    * Clients
+        * [CP-3882] - Add display for gender and KYC Failed fields in client profile
+    * External Services
+        * [CP-3932] - Revamp the notification external service UI
+## Version 1.4.2 - Community 1.0.0
+    * Maker Checker 
+        * [CP-3873] - KYC Maker Checker UI Improvements
+        * [CP-3911] - Apply KYC backend changes to align with the updated requirements.
+        * [CP-3927] - Enable search by client name and order by date on Failed KYC tab.
 
 ## Version 1.4.1 - Community 1.0.0